requirements.txt
rdflib==7.0.0
SPARQLWrapper==2.0.0
web3==6.15.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keys==0.4.0
eth-utils==2.2.0
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0